+++ Stability models/InitialSaturation.F90
@@ -47,7 +47,7 @@ subroutine InitialSaturation()
 			iLandUse = lulc(i,j)            
 !
 !           Check null
-            IF ((iZone .NE. nodata) .AND. (iLandUse .NE. nodata)) THEN
+            IF ((iZone .NE. INT(nodata)) .AND. (iLandUse .NE. INT(nodata))) THEN
 !            
                 Slope = slopeGrid(i,j)
                 Zmax = Gaussh(iZone)%mean
